+++ src/openeo_aggregator/constants.py
@@ -2,8 +2,10 @@ JOB_OPTION_SPLIT_STRATEGY = "split_strategy"
 
 JOB_OPTION_TILE_GRID = "tile_grid"
 
-# Experimental feature to force a certain upstream back-end through job options
-JOB_OPTION_FORCE_BACKEND = "_agg_force_backend"
+# Force a certain upstream back-end through job options
+JOB_OPTION_FORCE_BACKEND = "federation:force-backend"
+# TODO: remove support for legacy variant
+JOB_OPTION_FORCE_BACKEND_LEGACY = "_agg_force_backend"
 
 
 class CROSSBACKEND_GRAPH_SPLIT_METHOD: